Project options simplified (#317)
Eliminated the unconventional string-ish project setting, which was mainly used for debug purposes and was not clear or intuitive.

Instead, project_file is a simple True or False setting, controlling whether project files are read and written to disk.

Previous special-case values like reparse and retranslate are now separate settings. Fixed the logic so that they actually work, as a bonus.

SubtitleProject no longer takes an Options on initialisation, since it was only used to read the project setting (and unpacking some special case values as hidden settings), and it could mislead people into thinking that other options passed in to the project would have any effect on the project.

project_file is now exposed in the GUI, allowing users to translate subtitles without creating a project file if they so desire. 

Translations will now be saved directly if the project file is not enabled, rather than being a sort of side-effect (admittedly quite an important one!).

Replaced some anomalous getter functions with properties, for consistency with the rest of the project.

For context, this was some of the oldest code in the project, dating from when it was basically just a simple command-line tool for batching subtitles and sending them to gpt-3.5 requesting a translated version. A lot has changed since then, and this system looked too much like a fossil!

Given that the code was old, cryptic, written without much regard to maintenance and not really touched in a very long time this change is likely to introduce obscure bugs with certain workflows, but it needs to be done and leaves us in a better state going forwards.

* Simplify `project` configuration option - just a boolean indicating whether to use a persistent project file or not.
* Other specialised configurations are implemented as separate settings.
* Allow using the GUI without project files
* Documentation update
* Simplify needs saving/writing logic
* Implicitly enable using project file if initialised with a project file.
* Updated readme.md