** BREAKING CHANGE: ** Changed the install flow to register packages via pyproject.toml (#336)
** BREAKING CHANGE ** - you will need to re-run the install script or manually `pip install .` after syncing to this revision.

Changed how the project is installed so that the PySubtitle and GUI modules are treated as packages, meaning that the `sys.path` hack is no longer required for scripts to be able to see them.

This also allows specifying separate dependencies for each module in the .toml file, so that they can be installed as optional dependencies with pip - no need to pull down the massive PySide6 library if you only want to use the command line tools.

Streamlined the install scripts so that they use the optional dependencies flow and only need to run a single unified pip install, after collecting all the information about the type of install.

`requirements.txt` no longer exists.