Added support for SSA/SSA and VTT formats
Added support for multiple file formats to LLM-Subtrans (initially just introducing .ass/.ssa and .vtt support).

This is the culmination of a long-running project to move towards a format-agnostic architecture, removing dependencies on the srt module and reducing assumptions about the specific format of subtitles we need to handle.

Many changes have gone into this CL, completed in multiple phases.

Primarily, we introduce a SubtitleFormatRegistry that enables the supported formats to be registered and queried.

A new SSAFileHandler implements support for .ass/.ssa, using the pysubs2 library to handle reading and writing the format. SRT-specific logic was consolidated in SrtFileHandler.

A custom VTTFileHandler is provided, focusing on preserving metadata across translation to the same format.

File handlers are selected automatically based on file extension when loading or saving subtitles.

A fallback mechanism is provided to autodetect the format based on content, using pysub2's format detection system.

User is given the option to change the output format when creating a new translation project.
The output path for the translation file is now determined based on language and format.
Conversely, the output format of the subtitles is determined by any custom output path specified on the CLI.

GUI display updated to show additional details for .ass subtitles (just the line style).

Rationalised the flow for saving projects and/or translations, rather than having to manually keep track of what to save.

Unit tests were greatly expanded to handle the increased complexity.